refactor code file reference.slice
